+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `level-of-detail` sandbox to test out the three `LOD` object.
 - Created `DebugTextPanel` to handle displaying updating text to the screen.
 - Removed `GLStats` and converted it's usage to `DebugTextPanel`.
+- Added river water to `olympia-realtime-light-test` terrain.
 
 ## v0.1.6 - Feb 26, 2021